refactor: join bookings dao to courses dao
